Refactored output() to be a function and POSIX complient and Refactored tests.
Added pass through of cut and rm commands because they generally work the same on all shells and it would be a LOT of work to mock them out (for little to no benefit).

Removed isNotBash() from spec_helper because I changed the CUT so it works on all shells (so it's not needed). I also added return code 200 global.

Refactored output() to be a function. This required a LOT of changes. This was done because it encapsulates the code. I hadn't been doing it that way because I thought POSIX didn't support functions, but POSIX just doesn't support the 'function' keyword.

I refactored tests such that invalid input tests are in a different file. I did this because:
- Most of those tests are almost identical so to create more in the future, I can just make a copy of one of the existing ones as the base for those sorts of test in the future
- It splits up some of the largest test files
- It should make it easier to make sweeping changes to those tests in the future if need be

I refactored how test files are named so the file names match the 'Describe" titles.

Refactored backUp() and log() to use a main.sh to handle importing needed stuff so that code is more like output().

Refactored output() to be fully POSIX complaint. Also updated it so it returns more helpful log messages and more accurate return codes.

Refactored createHeaderFooter() so it no longer include a new line character at the end of it's stdout. This was done because dash removes the new line character.

Updated code coverage min percentage to reflect what it currently is for output() and createHeaderFooter().