Refactored function docs and added unit tests of checkRequiredOpts()
Added methods to spec_helper that check if test output contains a special character. This has been done because diffrent shells handle producing special characters to std in diffrent ways and there's no way (that I know of) to normalize that behavor in CUT.

Added unit tests of checkRequiredOpts().

Refactored all method docs in project to make it more clear how to use them as well as including an author section.

Added passthroughs of echo because it was needed for the special character checking methods added to spec_helper.